feat core_plugin: add some core functionalities

+++ src/index.ts
@@ -8,21 +8,25 @@ import { can_access } from './lib/utils.js';
 import readline from 'readline/promises';
 import { global_i18n } from './lib/i18n.js';
 import fs, { copyFile } from 'fs/promises';
+import * as util from 'util';
 
 global_log.put('Supervisor initialize...');
 process.chdir(path.dirname(new URL(import.meta.url).pathname));
 
-const rl = readline.createInterface({ input: process.stdin });
+const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
 const config = await (async () => {
   if (await can_access(config_path)) return await read_config();
   else if (
     (await can_access(default_config_path)) &&
-    (await can_access(config_path, fs.constants.W_OK))
+    (await can_access(path.dirname(config_path), fs.constants.W_OK))
   ) {
     const docp = await rl.question(
-      global_i18n.tr('Config not present, copy from the default config?')
+      global_i18n.tr('Config not present, copy from the default config? [Y(es)/n]')
     );
-    if (!docp) global_log.put('Denied by user, quitting...');
+    if (docp && docp.trim().match(/no?/i)) {
+      global_log.put('Denied by user, quitting...');
+      throw new Error(global_i18n.tr('Copying default config denied by the user'));
+    }
     await copyFile(default_config_path, config_path);
     return await read_config();
   } else throw new Error(global_i18n.tr("Can't read config or copy the default config"));
@@ -32,24 +36,29 @@ if (config.log_level) global_log.level = config.log_level;
 
 type CtrlMessage = SupervisorMessage | { kind: 'exit'; code: number } | { kind: 'watchdog' };
 
+global_log.put('Enter supervisor main loop...');
 while (true) {
   const cp = child_process.fork('./init.js');
   const subevent = await new Promise<CtrlMessage>((res) => {
     let watchdog_fed = false;
     cp.on('message', (message) => {
-      const msg = JSON.parse(String(message)) as SupervisorMessage;
+      const msg = message as SupervisorMessage;
+      if (!('kind' in msg))
+        throw new Error(`Unrecognized message: ${util.inspect(msg, { colors: true })}`);
       switch (msg.kind) {
         case 'reboot':
-          global_log.header('Received application reboot message, rebooting...');
+          global_log.put('Received application reboot message, rebooting...');
           res(msg);
           break;
         case 'shutdown':
-          global_log.header('Received application shutdown message, shutting down...');
+          global_log.put('Received application shutdown message, shutting down...');
           res(msg);
           break;
         case 'watchdog':
           watchdog_fed = true;
           break;
+        default:
+          throw new Error(`Unrecognized message: ${util.inspect(msg, { colors: true })}`);
       }
     });
 
@@ -72,6 +81,8 @@ while (true) {
       }, config.watchdog_interval * 2000);
     } else global_log.tmpl()`[Watchdog] Watchdog disabled, idle`;
   });
+  cp.kill('SIGKILL');
+  global_log.header_tmpl()`Subprocess killed with reason ${subevent}, code ${cp.exitCode}`;
   if (subevent.kind === 'watchdog') {
     if (!(config.debug?.reboot_on_watchdog ?? true)) break;
   } else if (subevent.kind !== 'shutdown') break;
